Multiple Iterations of calculation
The intersection of one array and its exact duplicate is 100%, derp.

Also, it turns out that the first iteration of the input may result in
ZERO duplicates, so this is set up to run it over and over until it
finds one.

However. It runs forever and doesn’t return, so that’s no good.